Add TypeConverter to collections
