+++ Source/HelixToolkit.Geometry/Collection/Vector3Collection.cs
@@ -1,9 +1,11 @@
-﻿using System.Globalization;
+﻿using System.ComponentModel;
+using System.Globalization;
 using System.Text;
 
 namespace HelixToolkit;
 
 [Serializable]
+[TypeConverter(typeof(Vector3CollectionConverter))]
 public sealed class Vector3Collection : FastList<Vector3>
 {
     public Vector3Collection()
