+++ Source/HelixToolkit.Geometry/TypeConverters/Vector2CollectionConverter.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel;
 using System.Globalization;
 
-namespace HelixToolkit.SharpDX.Utilities;
+namespace HelixToolkit;
 
 public sealed class Vector2CollectionConverter : FromToStringTypeConverter
 {
